requirements.txt
mysql-connector
mysql-connector-python
mysql-connector-python-rf